Prefabs

+++ TeamProject/GameObject.h
@@ -82,7 +82,7 @@ namespace NCL {
 
 			template<class T>
 			void AddComponent(Component * obj);
-
+			
 			template<class T>
 			void RemoveComponent()
 			{
@@ -155,7 +155,7 @@ namespace NCL {
 			void RemoveScript()
 			{
 
-				for (int i = 0; i < scripts.size(); i++)
+				for (int i = 0; i < (int)scripts.size(); i++)
 				{
 					if (dynamic_cast<T>(scripts[i]))
 					{
